refactor unstructured parser
