+++ examples/evaluation/document-search/basic/evaluate.py
@@ -53,7 +53,7 @@ config = {
             },
             "parsers": {
                 "txt": {
-                    "type": "ragbits.document_search.ingestion.parsers.unstructured:UnstructuredDefaultProvider",
+                    "type": "ragbits.document_search.ingestion.parsers.unstructured:UnstructuredDocumentParser",
                 },
             },
             "source": {
