--- packages/ragbits-document-search/src/ragbits/document_search/ingestion/parsers/unstructured/__init__.py
@@ -1,5 +1,0 @@
-from .default import UnstructuredDefaultProvider
-from .images import UnstructuredImageProvider
-from .pdf import UnstructuredPdfProvider
-
-__all__ = ["UnstructuredDefaultProvider", "UnstructuredImageProvider", "UnstructuredPdfProvider"]
