+++ packages/ragbits-document-search/tests/unit/test_document_parsers.py
@@ -7,37 +7,25 @@ from ragbits.core.utils.config_handling import ObjectContructionConfig
 from ragbits.document_search.documents.document import DocumentMeta, DocumentType
 from ragbits.document_search.ingestion.parsers.base import DocumentParser, DocumentTypeNotSupportedError
 from ragbits.document_search.ingestion.parsers.dummy import DummyProvider
-from ragbits.document_search.ingestion.parsers.unstructured.default import UnstructuredDefaultProvider
-from ragbits.document_search.ingestion.parsers.unstructured.images import UnstructuredImageProvider
-from ragbits.document_search.ingestion.parsers.unstructured.pdf import UnstructuredPdfProvider
+from ragbits.document_search.ingestion.parsers.unstructured import UnstructuredDocumentParser
 
 
-@pytest.mark.parametrize("document_type", UnstructuredDefaultProvider.SUPPORTED_DOCUMENT_TYPES)
+@pytest.mark.parametrize("document_type", UnstructuredDocumentParser.supported_document_types)
 def test_unsupported_provider_validates_supported_document_types_passes(document_type: DocumentType):
-    UnstructuredDefaultProvider().validate_document_type(document_type)
-
-
-@pytest.mark.parametrize("document_type", UnstructuredPdfProvider.SUPPORTED_DOCUMENT_TYPES)
-def test_unsupported_pdf_provider_validates_supported_document_types_passes(document_type: DocumentType):
-    UnstructuredPdfProvider().validate_document_type(document_type)
-
-
-@pytest.mark.parametrize("document_type", UnstructuredImageProvider.SUPPORTED_DOCUMENT_TYPES)
-def test_unsupported_images_provider_validates_supported_document_types_passes(document_type: DocumentType):
-    UnstructuredImageProvider().validate_document_type(document_type)
+    UnstructuredDocumentParser().validate_document_type(document_type)
 
 
 def test_unsupported_provider_validates_supported_document_types_fails():
     with pytest.raises(DocumentTypeNotSupportedError) as err:
-        UnstructuredDefaultProvider().validate_document_type(DocumentType.UNKNOWN)
+        UnstructuredDocumentParser().validate_document_type(DocumentType.UNKNOWN)
 
-    assert "Document type unknown is not supported by the UnstructuredDefaultProvider" in str(err.value)
+    assert "Document type unknown is not supported by the UnstructuredDocumentParser" in str(err.value)
 
 
 @patch.dict(os.environ, {}, clear=True)
 async def test_unstructured_provider_raises_value_error_when_api_key_not_set():
     with pytest.raises(ValueError) as err:
-        await UnstructuredDefaultProvider(use_api=True).parse(
+        await UnstructuredDocumentParser(use_api=True).parse(
             DocumentMeta.create_text_document_from_literal("Name of Peppa's brother is George.")
         )
 
@@ -47,7 +35,7 @@ async def test_unstructured_provider_raises_value_error_when_api_key_not_set():
 @patch.dict(os.environ, {}, clear=True)
 async def test_unstructured_provider_raises_value_error_when_server_url_not_set():
     with pytest.raises(ValueError) as err:
-        await UnstructuredDefaultProvider(api_key="api_key", use_api=True).parse(
+        await UnstructuredDocumentParser(api_key="api_key", use_api=True).parse(
             DocumentMeta.create_text_document_from_literal("Name of Peppa's brother is George.")
         )
